requirements.txt
# requirements.txt — E-Commerce Batch Analytics Pipeline v2
# Python 3.13 | Batch only (no Kafka/Spark)
# Install: pip install -r requirements.txt

# ── Core ──────────────────────────────────────────────────────────────────────
pandas==2.2.3
numpy==2.1.3
python-dotenv==1.0.1
python-dateutil==2.9.0
pytz==2024.2

# ── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36

# ── dbt (local dev — Python 3.13 compatible) ──────────────────────────────────
# NOTE: dbt-utils adalah dbt package, install via: cd dbt && dbt deps
# NOTE: Airflow container pakai dbt-core==1.7.4 (Python 3.8 compatible)
dbt-core==1.8.9
dbt-postgres==1.8.2

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-cov==6.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
click==8.1.7
rich==13.9.4
tabulate==0.9.0
